feat: Remove problematic kit loading feature
The "Load Kit" feature, which provided preset drum kits, has been removed. This is due to persistent and critical errors encountered during audio data decoding (specifically with `atob`) which led to an "Unable to decode audio data" state.

This feature will be re-evaluated for a future release with a more robust implementation to address the underlying audio decoding issues.